Update task that looks for existing packages (#1)
A task in `tasks/main.yml` is described as checking whether yaourt is
installed. This description is incorrect. The task actually checks to
see whether a user-supplied list of packages is installed. Fix this
error.

Let this task always run in normal mode, even when check mode is
enabled. It's safe for this task to do so, as this task's purpose is to
gather information, not to change the state of the target system. This
make this role behave better when check mode is enabled, as the variable
created by this task (`archlinux_aur_check_result`) will now be created.
Plus, the ansible-lint tool is happier with the result.